Create Sentence_transformers_cossim.py
Here, we first concat description and genres for each book and find the embeddings. The cosine similarity between the embeddings is used to make recommendations. Then again, we evaluate the model using two ways, ratings and overlapping genres and we get average precision of 0.62 using first evaluation method and a mean average precision of 0.33 using second evaluation method